Feat : use-status 라우트 page.tsx에 Buttons 컴포넌트 병합 및 주석 변경
- 예시에 대한 이해가 쉽도록 useFormStatus 훅을 사용한 컴포넌트를 라우트 파일 내부로 병합, 예시 작성
